==> code/1a_locations-for-requests.ipynb <==
# Create location for Bird API requests

## Setup

%run -i path.py
%run -i setup.py

## Functions

%run -i functions.py

# Download Turin networks
# Do not run this cell if you already downloaded the networks
G = ox.graph_from_place('Turin, Piemonte, Italy', network_type='all')

# Save to disk
# Do not run this cell if you already downloaded the networks
ox.save_graph_shapefile(G,PATH['data']+'graph_shapefile/Turin/')

# gdf of the edges of Turin street network
gdf_edges = gpd.read_file(PATH['data']+'graph_shapefile/Turin/edges.shp')

gdf_edges.plot(figsize=(20,10),alpha=0.2)

## Create a grid for the city center

# total area of the grid
ymax=45.10
ymin=45.03
xmin=7.62
xmax=7.72


# how many cells across and down
n_cells = 10

cell = create_grid(xmin,xmax,ymin,ymax,n_cells)

# this list contains the indices of peripheral boxes that we will remove
suburbs = [0,1,2,6,7,56,57,64,65,66,72,73,74,80,81,82,87]

# we will use the centroids of the boxes as coordinates for the query
# in the next cells we remove the peripheral points

points = cell.copy()

points['geometry'] = points['geometry'].centroid

points=points.drop(suburbs,axis=0)

print("Number of queries: ",len(points))
print("Side of the boxes: ",np.round(distance_meters(points['geometry'].loc[3], points['geometry'].loc[4])[1]),' m')

fig, ax = plt.subplots(nrows=1,ncols=1,figsize=(20,10),sharex=True, sharey=True)

ax = cell.plot(ax=ax,facecolor='none',edgecolor='black')
ax = gdf_edges.plot(ax = ax,alpha=0.2)
points.plot(ax=ax,figsize=(20,10), color='green')

## Create the .csv file with the boxe's centroids for the requests

# centroids

lat = points['geometry'].y
long = points['geometry'].x

lat_list = []
long_list = []
for i in range(len(lat)):
    lat_list.append(lat.iloc[i])
    long_list.append(long.iloc[i])
    
a = pd.DataFrame({'latitude':lat_list})
b = pd.DataFrame({'longitude':long_list})

c = pd.concat([a,b],axis=1)

# save the coordinates of these centroids
# we will use them to make the requests
c.to_csv(PATH['data']+ placeid + "/"+'request_location_centroids_turin.csv')
