==> code/2b_generate-ODdata.ipynb <==
# Generate OD data from e-scooter positions

In this notebook we process the e-scooter position data, collected using the notebook "1b_collect-escooter-data.ipynb" to generate Origin-Destination data that we can use to define a trips data set.

## Setup

%run -i path.py
%run -i setup.py

## Functions

%run -i functions.py

# create a single dataframe with the collected data

start = 500
end = 555

start_day = '2021-05-26'
end_day = '2021-10-28'

frames = []

for i in range(start,end):
    
    print(i+1,'/',end)
    
    # load the data
    frame = pd.read_csv(PATH['data_api']+ placeid + "/"+'TimeSlot'+str(i)+'_Start'+str(start_day)+'_End'+str(end_day)+'.csv')

    # delete the overlapping data
    frame = frame.drop_duplicates(['id'],ignore_index=True)
    
    # append to a list of df
    frames.append(frame)
    
    clear_output(wait=True)

# gdf of the edges of Turin street network
gdf_edges = gpd.read_file(PATH['data']+'graph_shapefile/Turin/edges.shp')

The function OD_data(frames,code,d_T,t_T) finds the positions of the vehicle identified by the variable "code" that are stored in the list of dataframes "frames". Then it identifies the movement made by the given vehicle. 2 threshold are defined to define the movements:
* "d_T" -> threshold (in meters) to define a movement: if the position of a vehicle changes of > d_T meters from frames[i-1] to frames[i], we consider it as a movement.
* "t_T" -> threshold (in minutes) to define a movement: we do not consider a change in the position as a movement if the time between the detection in frames[i-1] and frames[i] is > t_T minutes

It returns the coordinates of the Origin (O_lat and O_lon), the battery level of the Origin (O_battery), the coordinates of the Destination (D_lat and D_lon), the battery level of the Destination (D_battery), the codes that identifies the vehicle that made this movements (codes).

# find OD_data for TimeSlot from i = start to i = end

d_T = 100 # threshold (in meters)
t_T = 90 # threshold (in minutes)

OD_list = []

# we take 100 frames (not all together), so we do not miss different vehicles with same unique code
for f in range(100,len(frames),100):
    
    # find unique values for 'code' in all df, then create a list with all the 'code' unique values found in data collection
    unique_code = find_unique_code(frames[f-100:f])

    # remove the data in which 'code' refears to more than 1 e-scooter
    code_list, doubles = remove_doubles(frames[f-100:f],unique_code)
    
    O_lat_tot = []
    O_lon_tot = []
    O_time_tot = []
    O_battery_tot = []
    D_lat_tot = []
    D_lon_tot = []
    D_time_tot = []
    D_battery_tot = []
    codes_tot = []

    for j,code in enumerate(code_list):
    
        print(j+1,'/',len(code_list))
    
        O_lat,O_lon,O_time,O_battery,D_lat,D_lon,D_time,D_battery,codes = OD_data(frames[f-100:f],code,d_T,t_T)
    
        O_lat_tot = O_lat_tot + O_lat
        O_lon_tot = O_lon_tot + O_lon
        O_time_tot = O_time_tot + O_time
        O_battery_tot = O_battery_tot + O_battery
        D_lat_tot = D_lat_tot + D_lat
        D_lon_tot = D_lon_tot + D_lon
        D_time_tot = D_time_tot + D_time
        D_battery_tot = D_battery_tot + D_battery
        codes_tot = codes_tot + codes
    
        clear_output(wait=True)
        
    data = {
    'O_lat' : O_lat_tot,
    'O_lon' : O_lon_tot,
    'D_lat' : D_lat_tot,
    'D_lon' : D_lon_tot,
    'O_time' : O_time_tot,
    'D_time' : D_time_tot,
    'O_battery' : O_battery_tot,
    'D_battery' : D_battery_tot,
    'code' : codes_tot
    }
    
    # set the od_data into a df
    OD_matrix0_100 = pd.DataFrame(data, columns = list(data.keys()))

    #sort dataframe by O_time
    OD_matrix0_100 = OD_matrix0_100.sort_values(by='O_time')
    OD_matrix0_100 = OD_matrix0_100.reset_index()
    OD_matrix0_100 = OD_matrix0_100.drop(['index'],axis=1)

    # add a column with O_battery - D_battery: we use this value to remove the movements made by the company (negative values)
    OD_matrix0_100['batt_diff'] = OD_matrix0_100['O_battery']-OD_matrix0_100['D_battery']
      
    # remove movements made by the company
    a = OD_matrix0_100[OD_matrix0_100['D_battery'] < 99] # remove data if the destination battery level is > 98 (it means it did not decrease after a trip)
    clean_OD = a[a['batt_diff']>0].reset_index() # remove data if battery_origin-battery_destination < 1 
    clean_OD = clean_OD.drop(['index'],axis=1)
    
    OD_list.append(clean_OD)

final_OD = pd.concat(OD_list,axis=0)

# export the OD data
final_OD.to_csv(PATH['data']+ placeid + "/" + "OD_data.csv")
